+++ src/Type/InputMediaVenue.php
@@ -1,0 +1,49 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Phptg\BotApi\Type;
+
+use Phptg\BotApi\FileCollector;
+
+/**
+ * @see https://core.telegram.org/bots/api#inputmediavenue
+ *
+ * @api
+ */
+final readonly class InputMediaVenue implements InputMedia
+{
+    public function __construct(
+        public float $latitude,
+        public float $longitude,
+        public string $title,
+        public string $address,
+        public ?string $foursquareId = null,
+        public ?string $foursquareType = null,
+        public ?string $googlePlaceId = null,
+        public ?string $googlePlaceType = null,
+    ) {}
+
+    public function getType(): string
+    {
+        return 'venue';
+    }
+
+    public function toRequestArray(?FileCollector $fileCollector = null): array
+    {
+        return array_filter(
+            [
+                'type' => $this->getType(),
+                'latitude' => $this->latitude,
+                'longitude' => $this->longitude,
+                'title' => $this->title,
+                'address' => $this->address,
+                'foursquare_id' => $this->foursquareId,
+                'foursquare_type' => $this->foursquareType,
+                'google_place_id' => $this->googlePlaceId,
+                'google_place_type' => $this->googlePlaceType,
+            ],
+            static fn(mixed $value): bool => $value !== null,
+        );
+    }
+}